p4i integration Integrate 116529 from 3 to main: Integrate changes from 3.2
[git-p4: depot-paths = "//depot/qt/main/": change = 116530]

+++ examples/xml/tagreader/structureparser.cpp
@@ -32,6 +32,6 @@ bool StructureParser::startElement( const QString&, const QString&,
 
 bool StructureParser::endElement( const QString&, const QString&, const QString& )
 {
-    indent.remove( 0, 4 );
+    indent.remove( (uint)0, 4 );
     return TRUE;
 }
